Revert "Add some sample files so linting can be tested"
This reverts commit 173af2a7b9d9a0c2ff77dadeeb2471b5e410137c.
