--- dependencies.d.ts
@@ -1,5 +1,0 @@
-declare module 'loglevel' {
-  function debug(message: string | Record<string, unknown>): void;
-  const log = { debug };
-  export default log;
-}
